// src/app/models/servicio.ts
export class Servicio {
    id?: any;
    nombre?: any;
    estado?: any;
    codigo_barra?: any;
    precioCosto?: any;
    iva?: any;
    precioIva?: any;
    precioStandar?: any;
    margenProduccion?: any;
    precioProduccion?: any;
    margenVenta?: any;
    precioVenta?: any;
    precioFinal?: any;
    nombreEstado?: any;
}